Rework install/unlinstall flow
Restructure the code and use the same approach as the launcher tmd restorer to deal with tmds.
Store all the known tmds for all the launcher versions with their corresponding shas, and use them when restoring the console.
Now if the main tmd isn't either the stock one or one with the patch done by this program, it's treated a invalid and will be restored from the saved ones when uninstalling.
This breaks for now proto/dev consoles as there is no handling for "default tmds" on them and the program logic has to be changed accordingly